+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - button of opening right sidebar
+- displaying of panel list 
 
 ## [0.9.0]